requirements.txt
Django~=4.2.1
python-dotenv~=1.0.0
djangorestframework~=3.14.0